LayoutTests/fast/events/ios/dragstart-hang-does-not-dismiss-context-menu.html

<!DOCTYPE HTML><!-- webkit-test-runner [ dragInteractionPolicy=always-enable useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<head>
<style>
body, html {
    margin: 0;
}
</style>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("This test requires WebKitTestRunner. To manually test, load the page in a configuration where drag and drop is enabled by default; long press the image to show the context menu, and verify that the context menu remains visible");

    startedDrag = false;

    let image = document.querySelector("img");
    image.addEventListener("dragstart", () => {
        const startTime = Date.now();
        while (true) {
            if (Date.now() - startTime > 2500)
                break;
        }
        getSelection().setPosition(document.body, 1);
        startedDrag = true;
    });

    if (!window.testRunner)
        return;

    await UIHelper.longPressElement(image);
    await UIHelper.waitForContextMenuToShow();
    await shouldBecomeEqual("startedDrag", "true");
    await UIHelper.delayFor(1000);

    previewBounds = await UIHelper.contextMenuPreviewRect();
    shouldBeGreaterThanOrEqual("previewBounds.width", "0");
    shouldBeGreaterThanOrEqual("previewBounds.height", "0");

    await UIHelper.activateAt(1, 1);
    await UIHelper.waitForContextMenuToHide();

    finishJSTest();
});
</script>
</head>
<body>
    <img src="../resources/abe.png">
    <p id="description"></p>
    <p id="console"></p>
</body>
</html>
